refactor: :zap: Bundler migration to vite Global website refactor:  - Migration from webpack to vite  - Removal of deprecated dependancies  - Refactor of the code (sheet and loading)  - Include alphaTab as standalone (see #22)
