--- README.md
@@ -1,63 +1,0 @@
-# Tablatures
-
-<a href="https://github.com/mlhoutel/Tablatures/blob/main/.github/workflows/main.yml" alt="Build">
-  <img src="https://img.shields.io/github/workflow/status/mlhoutel/Tablatures/Build%20Vue/main?style=flat-square" />
-</a>
-
-<a href="https://github.com/mlhoutel/Tablatures/blob/main/package.json" alt="Version">
-  <img src="https://img.shields.io/github/package-json/v/mlhoutel/Tablatures?style=flat-square" />
-</a>
-
-<a href="https://github.com/mlhoutel/Tablatures/search?l=vue" alt="Language">
-  <img src="https://img.shields.io/github/languages/top/mlhoutel/Tablatures?style=flat-square" />
-</a>
-
-<a href="https://github.com/mlhoutel/Tablatures/projects/1" alt="Roadmap">
-  <img src="https://img.shields.io/badge/roadmap-available-brightgreen?style=flat-square" />
-</a>
-
-<a href="https://github.com/mlhoutel/Tablatures/blob/main/LICENSE/" alt="License">
-  <img src="https://img.shields.io/github/license/mlhoutel/Tablatures?style=flat-square" />
-</a>
-
-
-<img src="./src/assets/imgs/Tablatures.png">
-
-## Functionnalities
-
-Read and play tabs quickly
-
-- **Import a music score** (.gp3, .gp4, .gp5, .gpx, .gp, .xml, .cap or .tex)
-- **Browse a music score database** (IN DEV)
-- **Browse documentation** (IN DEV)
-
-## Using
-
-- VueJs: https://github.com/vuejs/vue
-- Vuetify: https://github.com/vuetifyjs/vuetify
-- AlphaTab: https://github.com/CoderLine/alphaTab
-
-## Project setup
-
-```
-npm install
-```
-
-### Compiles and hot-reloads for development
-
-```
-npm run serve
-```
-
-### Compiles and minifies for production
-
-```
-npm run build
-```
-
-### Lints and fixes files
-
-```
-npm run lint
-```
-
