+++ src/components/AppFooter.vue
@@ -1,11 +1,11 @@
 <template>
-  <v-footer absolute padless color="blue darken-3" dark>
+  <v-footer color="blue darken-3" dark>
     <v-row justify="center" no-gutters>
       <v-col class="text-center pa-2" cols="12"><b>tablatures</b> - an open source project for <b>guitar players</b> </v-col>
     </v-row>
   </v-footer>
 </template>
-<script lang="ts">
+ <script>
 import Vue from "vue"
 
 export default Vue.extend({
