--- src/raw-loader.d.ts
@@ -1,4 +1,0 @@
-declare module "!!raw-loader!*" {
-  const contents: string
-  export = contents
-}
