--- src/shims-vuetify.d.ts
@@ -1,4 +1,0 @@
-declare module "vuetify/lib/framework" {
-  import Vuetify from "vuetify";
-  export default Vuetify;
-}
