+++ src/views/Search.vue
@@ -10,7 +10,7 @@
   </v-container>
 </template>
 
-<script lang="ts">
+ <script>
 import Vue from "vue"
 
 export default Vue.extend({
