--- vue.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  transpileDependencies: ["vuetify"],
-  publicPath: process.env.NODE_ENV === "production" ? "/Tablatures/" : "/",
-  devServer: {
-    disableHostCheck: true,
-    port: 8080,
-    public: "0.0.0.0:8080",
-  },
-  configureWebpack: (config) => {
-    config.output.globalObject = "this"
-  },
-}
